LayoutTests/http/tests/security/mixedContent/insecure-css-in-main-frame.html

<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}

window.data = null;
window.addEventListener("message", function (e) {
    window.data = e.data;
    shouldBe("window.data", "'blocked'");

    if (window.testRunner)
        testRunner.notifyDone();
}, false);

</script>
<p>This test opens a window that loads an insecure style sheet.  We should
trigger a mixed content callback because the main frame in the window is HTTPS
but is running insecure content.</p>
<script>
onload = function() {
    window.open("https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-css.html");
}
</script>
</body>
</html>
